requirements.txt
AutoFeatSelect==0.1.5
hydra-core==1.3.2
imbalanced_learn==0.12.3
matplotlib==3.9.1.post1
numpy==2.0.1
omegaconf==2.3.0
optuna==3.6.1
pandas==2.2.2
psutil==6.0.0
PyYAML==6.0.2
scikit_learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
shap==0.46.0
statsmodels==0.14.2
torch==2.4.0
tqdm==4.66.5
xgboost==2.1.1
lightgbm==4.5.0
